Added comments and changed build summary to detail

+++ ccb_rollback_build.sql
@@ -5,27 +5,12 @@ create or alter procedure ccb.rollback_build (@build_id int) as
 begin
 
 set nocount on
-/*
-select *
-into DataStage.ccb.PlanSettingValue_BACKUP
-from CAV22.config.PlanSettingValue
-where PlanId = 18722
 
-select *
-into DataStage.ccb.TreatmentPlanSettingValue_BACKUP
-from CAV22.config.TreatmentPlanSettingValue
-where PlanId = 18722
+begin tran
 
-select a.*
-into DataStage.ccb.IncentiveAmounts_BACKUP
-from CAV22.dbo.IncentiveAmounts a
-join CAV22.dbo.IncentiveTiers b
-on a.IncentiveTier_Id = b.Id
-where b.Plan_Id = 18722
-*/
-
-
-drop table if exists #Build_Summary
+declare @commit_deploy bit = 1
+-- Grabbing ALL records for the chosen ConfigChangeBuildID from ccb.ConfigChangeBuildDetail
+drop table if exists #Build_Detail
 select bls.EntityTypeID
      , ent.EntityTypeName
 	 , bls.EntityID
@@ -45,9 +30,11 @@ select bls.EntityTypeID
 	        when cop.ChangeOperationName = 'Delete' then 'Insert'
 			else cop.ChangeOperationName
 		end as ChangeOperationName
-	 , bls.Ticket
-into #Build_Summary
-from ccb.ConfigChangeBuildSummary bls
+	 , ccb.Ticket
+into #Build_Detail
+from ccb.ConfigChangeBuildDetail bls
+     inner join ccb.ConfigChangeBuild ccb
+	         on ccb.ConfigChangeBuildID = bls.ConfigChangeBuildID
      inner join ccb.Config cfg
 	         on cfg.ConfigID = bls.ConfigID
 	 inner join ccb.ChangeOperation cop
@@ -56,24 +43,31 @@ from ccb.ConfigChangeBuildSummary bls
 	         on ent.EntityTypeID = bls.EntityTypeID
 	 left  join CAV22.dbo.TreatmentCodes tmc
 	         on bls.ProcedureID = (case when len(tmc.Code) = 5 then concat(9, tmc.Code) else tmc.Code end)
-where ConfigChangeBuildID = @build_id
+where bls.ConfigChangeBuildID = @build_id
 
+
+---------------------------------------------------------------------------------------------------------------
+------------------------------------------ CONFIGURATION CHANGES ----------------------------------------------
+---------------------------------------------------------------------------------------------------------------
+
+-- Grabbing build detail records pertaining to configurations and adding setting definitions necessary for config.Upsert procedures
 drop table if exists #Config_Upserts
 select EntityTypeName
      , EntityID
 	 , case when EntityTypeName in ('client', 'treatment_client') then cln.VitalsClientId else null end as VitalsClientId
      , case when EntityTypeName in ('plan', 'treatment_plan') then EntityID else null end as PlanId
-	 , TreatmentCode
+	 , bls.TreatmentCode
 	 , def.[Name] as SettingDefinitionName
 	 , def.SettingGroupKey
 	 , ctg.[Name] as SettingCategoryName
 	 , ConfigValueNew as SettingValue
-	 , ConfigValueOld as SettingValue_OLD
+	 , ConfigValueOld as SettingValue_Old_Build
+	 , cast(coalesce(tpv.SettingValue, tcv.SettingValue, psv.SettingValue, csv.SettingValue) as varchar) as SettingValue_Old_Actual
 	 , Ticket as ModifiedReason
 	 , ROW_NUMBER() over (order by EntityTypeName, EntityID, TreatmentCode, SettingGroupKey, def.[Name]) as config_step
 	 , 0 as config_step_complete
 into #Config_Upserts
-from #Build_Summary bls
+from #Build_Detail bls
      inner join CAV22.config.SettingDefinition def
 	         on def.[Name] = bls.ConfigName
 	 inner join CAV22.config.SettingCategory ctg
@@ -81,9 +75,30 @@ from #Build_Summary bls
 	 left  join CAV22.dbo.Clients cln
 	         on bls.EntityID = cln.Id
 			and bls.EntityTypeName in ('client', 'treatment_client')
+	 left  join CAV22.config.ClientSettingValue csv
+	         on csv.ClientId = bls.EntityID
+			and bls.EntityTypeName = 'client'
+			and csv.SettingDefinitionId = def.Id
+	 left  join CAV22.config.PlanSettingValue psv
+	         on psv.PlanId = bls.EntityID
+			and bls.EntityTypeName = 'plan'
+			and psv.SettingDefinitionId = def.Id
+	 left  join CAV22.config.TreatmentClientSettingValue tcv
+	         on tcv.ClientId = bls.EntityID
+			and bls.EntityTypeName = 'treatment_client'
+			and tcv.SettingDefinitionId = def.Id
+			and bls.ProcedureID = (case when len(tcv.TreatmentCode) = 5 then concat(9, tcv.TreatmentCode) else tcv.TreatmentCode end)
+	 left  join CAV22.config.TreatmentPlanSettingValue tpv
+	         on tpv.PlanId = bls.EntityID
+			and bls.EntityTypeName = 'treatment_plan'
+			and tpv.SettingDefinitionId = def.Id
+			and bls.ProcedureID = (case when len(tpv.TreatmentCode) = 5 then concat(9, tpv.TreatmentCode) else tpv.TreatmentCode end)
 where ConfigGroup = 'configuration'
 
---select * from DataStage.ccb.Config_Upsert_Test_BACKUP
+-- if the current actual Setting Values do not match the old values in build detail, rollback
+if (select count(*) from #Config_Upserts where SettingValue_Old_Build <> SettingValue_Old_Actual) > 0
+    set @commit_deploy = 0
+    goto End_Procedure
 
 declare @config_step int
 declare @config_step_max int
@@ -100,6 +115,7 @@ declare @modified_reason varchar(255)
 set @config_step = (select min(config_step) from #Config_Upserts)
 set @config_step_max = (select max(config_step) from #Config_Upserts)
 
+-- This loop updates configurations via the config.Upsert procedures using the transformed configuration records from the build detail
 while @config_step <= @config_step_max
 
 begin
@@ -165,11 +181,11 @@ begin
 	    set @config_step += 1
 end
 
---delete from CAV22.dbo.IncentiveAmounts where IncentiveTier_Id in (select Id from CAV22.dbo.IncentiveTiers where Plan_Id = 18722)
---insert into CAV22.dbo.IncentiveAmounts (Amount, NominalAmount, Active, IncentiveTier_Id, Procedure_Id, ModifiedReason)
---select Amount, NominalAmount, Active, IncentiveTier_Id, Procedure_Id, ModifiedReason
---from DataStage.ccb.IncentiveAmounts_BACKUP
+---------------------------------------------------------------------------------------------------------------
+----------------------------------------- INCENTIVE AMOUNTS CHANGES -------------------------------------------
+---------------------------------------------------------------------------------------------------------------
 
+-- Grabbing records from the build detail pertaining to Incentive Amounts and transforming them to fit into IncentiveAmounts
 drop table if exists #Incentive_Amount_Upserts
 select ChangeOperationID
 	 , ChangeOperationName
@@ -177,14 +193,17 @@ select ChangeOperationID
 	 , EntityName
      , ica.Id as IncentiveAmounts_Id
      , cast(ConfigValueNew as decimal(18,2)) as Amount
+	 , cast(ConfigValueOld as decimal(18,2)) as Amount_Old_Current
+	 , ica.Amount as Amount_Old_Actual
 	 , bls.ProcedureID as Procedure_Id
 	 , ict.Id as IncentiveTier_Id
 	 , 0 as incentive_step_complete
-	 , case when ica.ModifiedReason is not null then concat(ica.ModifiedReason, '; ', Ticket, ' Rollback') else concat(Ticket, ' Rollback') end as ModifiedReason
+	 , case when ica.ModifiedReason is not null then concat(ica.ModifiedReason, '; ', Ticket) else Ticket end as ModifiedReason
 into #Incentive_Amount_Upserts
-from #Build_Summary bls
+from #Build_Detail bls
      left  join CAV22.dbo.IncentiveTiers ict
 	         on ict.Plan_Id = bls.EntityID
+		        and ict.IsActive = 1
 			and ict.TierNumber = (case when ConfigName = 'static_tier_1' then 1
 			                           when ConfigName = 'static_tier_2' then 2
 									   when ConfigName = 'static_tier_3' then 3
@@ -195,17 +214,25 @@ from #Build_Summary bls
 where ConfigGroup = 'incentive_amounts'
   and EntityTypeName = 'treatment_plan'
 
+-- if the current actual Incentive Amounts do not match the old values in build detail, rollback
+if (select count(*) from #Incentive_Amount_Upserts where Amount_Old_Actual <> Amount_Old_Current) > 0
+    set @commit_deploy = 0
+    goto End_Procedure
+
+-- Deleting IncentiveAmount records 
 delete from CAV22.dbo.IncentiveAmounts
 where Id in (
 select IncentiveAmounts_Id
 from #Incentive_Amount_Upserts
 where ChangeOperationName = 'delete')
 
+-- Adding IncentiveAmount records
 insert into CAV22.dbo.IncentiveAmounts (Amount, NominalAmount, Active, IncentiveTier_Id, Procedure_Id, AddedBy, ModifiedReason, DateAdded)
 select Amount, 0.00 as NominalAmount, 1 as Active, IncentiveTier_Id, Procedure_Id, user as AddedBy, ModifiedReason, getdate() as DateAdded
 from #Incentive_Amount_Upserts
 where ChangeOperationName = 'insert'
 
+-- Updating IncentiveAmount records
 update ica
 set ica.Amount = upd.Amount
   , ica.DateModified = getdate()
@@ -220,7 +247,15 @@ update ccb.ConfigChangeBuild
 set IsDeployed = 0, DateLastRolledBack = getdate()
 where ConfigChangeBuildID = @build_id
 
-drop table if exists #Build_Summary
+begin
+    End_Procedure:
+    if @commit_deploy = 0
+	    rollback
+	else
+	    commit
+end
+
+drop table if exists #Build_Detail
 drop table if exists #Config_Upserts
 drop table if exists #Incentive_Amount_Upserts
